app/requirements.txt
alembic==1.0.0
amqp==2.3.2
billiard==3.5.0.4
celery==4.2.1
certifi==2018.8.13
chardet==3.0.4
click==6.7
Flask==1.0.2
Flask-Migrate==2.2.1
flask-simplelogin==0.0.6
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
future==0.16.0
idna==2.7
itsdangerous==0.24
Jinja2==2.10
kombu==4.2.1
Mako==1.0.7
MarkupSafe==1.0
oauthlib==2.1.0
pkg-resources==0.0.0
python-dateutil==2.7.3
python-editor==1.0.3
python-twitter==3.4.2
pytz==2018.5
redis==2.10.6
requests==2.19.1
requests-oauthlib==1.0.0
six==1.11.0
SQLAlchemy==1.2.11
urllib3==1.23
vine==1.1.4
Werkzeug==0.14.1
WTForms==2.2.1
